LPS-127654 Moves to content.compatibility.converter package
