+++ modules/apps/journal/journal-test/src/testIntegration/java/com/liferay/journal/content/compatibility/converter/test/JournalContentCompatibilityLayerTest.java
@@ -12,10 +12,10 @@
  * details.
  */
 
-package com.liferay.journal.util.test;
+package com.liferay.journal.content.compatibility.converter.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-import com.liferay.journal.util.JournalContentCompatibilityLayer;
+import com.liferay.journal.content.compatibility.converter.JournalContentCompatibilityLayer;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.xml.Document;
